+++ gui/__init__.py
@@ -1,0 +1,5 @@
+"""
+PathManager GUI Module
+
+PyQt6-based graphical user interface for PathManager.
+"""