+++ gui/dialogs/__init__.py
@@ -1,0 +1,5 @@
+"""
+PathManager GUI Dialogs
+
+Dialog windows for the PathManager GUI.
+"""