+++ gui/widgets/__init__.py
@@ -1,0 +1,5 @@
+"""
+PathManager GUI Widgets
+
+Custom PyQt6 widgets for the PathManager GUI.
+"""